delete 4 files, update 43 files and create 17 files

+++ app/Http/Controllers/LikeController.php
@@ -17,23 +17,17 @@ class LikeController extends Controller {
                   ->where('tweet_id', $request->tweet_id)
                   ->where('tweet_id', '!=', null)
                   ->first();
-      if ($like === null) {
-        Like::create($request->only(['user_id', 'tweet_id']));
-      } else {
-        $like->delete();
-      }
+      if ($like !== null) $like->delete();
+      else Like::create($request->only(['user_id', 'tweet_id']));
+
     } elseif ($request->isComment) {
 
       $like = Like::where('user_id', $request->user_id)
                   ->where('comment_id', $request->comment_id)
                   ->where('comment_id', '!=', null)
                   ->first();
-//      dd($like->toArray());
-      if ($like === null) {
-        Like::create($request->only(['user_id', 'comment_id']));
-      } else {
-        $like->delete();
-      }
+      if ($like !== null) $like->delete();
+      else Like::create($request->only(['user_id', 'comment_id']));
     }
 
   }
